API

NonparametricVI.compute_metric — Method

compute_metric(
    metric::KernelizedSteinDiscrepancy,
    pc::ParticleContainer,
    ρ;
    ad_backend
)

Compute the Kernelized Stein Discrepancy (KSD) between the particles in the ParticleContainer and the target log-density.

Arguments

metric::KernelizedSteinDiscrepancy: A KernelizedSteinDiscrepancy object that defines the kernel and other parameters used for KSD computation.

pc::ParticleContainer: The particle container holding the current set of particles.

ρ: A LogDensityProblem representing the target distribution's log-density function.

Keyword Arguments

ad_backend: The automatic differentiation backend to use for computing gradients required by the KSD.

Returns

The computed Kernelized Stein Discrepancy (KSD) value, a scalar representing the discrepancy between the particle distribution and the target distribution.

Details

This function calculates the KSD, a measure of the discrepancy between the distribution represented by the particles in pc and the target distribution defined by ρ. It utilizes the kernel specified in the metric object and the provided automatic differentiation backend ad_backend to compute the necessary gradients.

The function extracts the particle positions from pc.P and calls the kernelized_stein_discrepancy function to perform the KSD computation.

This function serves as a metric tracking tool during particle-based inference, allowing monitoring of the convergence of the particle distribution to the target.

NonparametricVI.constrained_particles — Method

constrained_particles(pc::ParticleContainer, model::DynamicPPL.Model)

Tranforming the position of samples in pc back to the constrained space

Arguments

pc::ParticleContainer: The container holding the unconstrained positions of the particles.

model::DynamicPPL.Model: The target Turing program

Returns

samples::Vector{<:DynamicPPL.VarInfo}: A vector of VarInfo objects, where each VarInfo contains a sample from the constrained parameter space of the model, corresponding to a particle in the ParticleContainer.

NonparametricVI.get_samples — Method

get_samples(pc::ParticleContainer, state::SVGDInferenceState)

Extract constrained samples from the particle container for Turing models

Arguments

pc::ParticleContainer: The particle container holding the current set of particles.

state::SVGDInferenceState: The internal state for the SVGD algorithm

Returns

Vector{DynamicPPL.VarInfo} a vector of varinfo containing parameters in the constrained space

NonparametricVI.infer! — Method

infer!(
    pc::ParticleContainer,
    state::SVGDInferenceState;
    iters::Integer,
    ad_backend=ADTypes.AutoForwardDiff(),
    verbose::Bool=false,
    track::Dict{String, Metric}=Dict()
)

Perform inference using Stein Variational Gradient Descent (SVGD).

Arguments

pc::ParticleContainer: The particle container holding the current set of particles.

state::SVGDInferenceState: The internal state object for the SVGD algorithm

Keyword Arguments

iters::Integer: The number of SVGD iterations to perform.

ad_backend=ADTypes.AutoForwardDiff(): The automatic differentiation backend to use for computing gradients. Defaults to ADTypes.AutoForwardDiff().

verbose::Bool=false: A boolean flag indicating whether to print progress information during the inference. Defaults to false.

Returns

SVGDInferenceReport: An SVGDInferenceReport object

Details

This function modifies the pc in-place, updating the positions of the particles.

NonparametricVI.init — Method

init(
    ρ,
    dynamics::ParticleDynamics;
    particle_initializer=NormalInitializer(),
    n_particles::Integer,
    ad_backend=ADTypes.AutoForwardDiff()
)

Initialize the particle container and the internal state for a particle-based inference algorithm, given a log-density problem.

Arguments

ρ: A LogDensityProblem representing the target distribution's log-density function.

dynamics::ParticleDynamics: The particle dynamics object that governs how particles evolve.

Keyword Arguments

particle_initializer: An object that initializes the positions of the particles. Defaults to NormalInitializer().

n_particles::Integer: The number of particles to initialize.

ad_backend=ADTypes.AutoForwardDiff(): The automatic differentiation backend to use if the provided ρ is not differentiable. Defaults to ADTypes.AutoForwardDiff().

Returns

pc: The initialized particle container, holding the initial positions of all particles.

state: The initialized internal state associated with the provided dynamics and ρ.

NonparametricVI.init — Method

init(
    mode::DynamicPPL.Model,
    dynamics::ParticleDynamics;
    particle_initializer=NormalInitializer(),
    n_particles::Integer,
    ad_backend=ADTypes.AutoForwardDiff()
)

Initialize the particle container and the internal state for a particle-based inference algorithm, given a log-density problem.

Arguments

ρ: A LogDensityProblem representing the target distribution's log-density function.

dynamics::ParticleDynamics: The particle dynamics object that governs how particles evolve.

Keyword Arguments

particle_initializer: An object that initializes the positions of the particles. Defaults to NormalInitializer().

n_particles::Integer: The number of particles to initialize.

ad_backend=ADTypes.AutoForwardDiff(): The automatic differentiation backend to use if the provided ρ is not differentiable. Defaults to ADTypes.AutoForwardDiff().

Returns

pc: The initialized particle container, holding the initial positions of all particles.

state: The initialized internal state associated with the provided dynamics and ρ.

NonparametricVI.init_state — Method

init_state(ρ, dynamics::SVGD, model::DynamicPPL.Model)

Initialize the SVGDInferenceState for using SVGD dynamics on Turing models

Arguments

ρ: A LogDensityProblem representing the target distribution.

dynamics::SVGD: An SVGD object defining the dynamics used to update the particles.

model::DynamicPPL.Model: The probabilistic model associated with the inference.

Returns

An SVGDInferenceState object initialized with the provided arguments.

NonparametricVI.init_state — Method

init_state(ρ, dynamics::SVGD)

Initialize the SVGDInferenceState for using SVGD dynamics with LogDensityProblem ρ

Arguments

ρ: A LogDensityProblem representing the target distribution's log-density function.

dynamics::SVGD: An SVGD object defining the dynamics used to update the particles.

Returns

An SVGDInferenceState object initialized with the provided arguments.

NonparametricVI.kernel_and_gradient_fn — Method

kernel_and_gradient_fn(K::KernelFunctions.Kernel, ad_backend)

Returns a function that computes the kernel value and its gradient with respect to the first argument.

Arguments

K::KernelFunctions.Kernel: The kernel function from KernelFunctions.jl

ad_backend: The automatic differentiation backend to use (e.g., AbstractDifferentiation.ForwardDiffBackend()).

Returns

A function k_∇k(x, a) that takes two arguments x and a (of compatible types for the kernel K) and returns a tuple containing:

The kernel value K(x, a).

The gradient of the kernel with respect to x, evaluated at x.

NonparametricVI.kernelized_stein_discrepancy — Method

kernelized_stein_discrepancy(P, q, K::KernelFunctions.Kernel; samplesize, ad_backend)

Computes the Kernelized Stein Discrepancy (KSD) between a set of samples P and a distribution q.

The KSD measures the discrepancy between two probability distributions by evaluating the expectation of two Stein operators applied to a kernel function.

Arguments

P: A matrix of samples from the empirical distribution. Each column represents a sample.

q: A LogDensityProblems.LogDensityProblem representing the target distribution.

K: A kernel function from KernelFunctions.Kernel.

samplesize: The number of sampled particles to evaluate KSD

ad_backend: An automatic differentiation backend from DifferentiationInterface.

Returns

The Kernelized Stein Discrepancy (KSD) as a scalar value.

Details

The function calculates the KSD using the following formula:

\[\text{KSD}(P, q) = \frac{1}{n(n-1)} \sum_{i=1}^n \sum_{j=1}^n u(P_i, P_j)\]

\[u(x, y) = \nabla s(x)^T k(x, y) \nabla s(y) + \nabla s(x)^T \nabla_y k(x, y) + \nabla_x k(x, y)^T \nabla s(y) + \text{tr}(\nabla_{x,y} k(x, y))\]

For more details see :

A Kernelized Stein Discrepancy for Goodness-of-fit Tests, Qiang Liu, Jason Lee, Michael Jordan

NonparametricVI.logdensityproblem_from_turing — Method

logdensityproblem_from_turing(model::DynamicPPL.Model, ad_backend::ADTypes.AbstractADType)

Constructs a differentiable LogDensityProblem from a Turing DynamicPPL.Model

This function takes a Turing model and an automatic differentiation backend, performs necessary transformations to the model's parameter space, and returns a LogDensityProblem that can be used with optimization or sampling algorithms that require gradient information.

Arguments

model::DynamicPPL.Model: The Turing probabilistic model.

ad_backend::ADTypes.AbstractADType: The automatic differentiation backend to use (e.g., ADTypes.AutoForwardDiff(), ADTypes.AutoZygote()).

Returns

∇ρ::LogDensityProblemsAD.ADgradient: A LogDensityProblem object that wraps the log-density function of the Turing model and provides gradient computation capabilities using the specified ad_backend. The parameters of this problem are in the unconstrained space.

NonparametricVI.particle_velocity — Method

particle_velocity(pc::ParticleContainer, ρ, pi, k_∇k, dynamics::SVGD)

Computes the velocity of a single particle based on the Stein Variational Gradient Descent (SVGD) update rule, potentially using a mini-batch of other particles.

Arguments

pc::ParticleContainer: The container holding the particles.

ρ: The log-density function (a LogDensityProblem) which must be differentiable.

pi::Int: The index of the particle for which to compute the velocity.

k_∇k: A function that takes two particle positions (as vectors) and returns a tuple containing the kernel value and the gradient of the kernel with respect to the first argument. This is generated by kernel_and_gradient_fn.

dynamics::SVGD: The SVGD dynamics object containing the kernel, step size, and batch size.

Returns

velocity::Vector: The computed velocity vector for the particle with index pi.

NonparametricVI.update_particles! — Method

update_particles!(ρ, pc::ParticleContainer, dynamics::SVGD)

Updates the positions of all particles in the ParticleContainer according to the Stein Variational Gradient Descent (SVGD) update rule.

Arguments

ρ: The log-density function (a LogDensityProblem) that the particles aim to sample from.

pc::ParticleContainer: The container holding the current positions of the particles. The particle positions are updated in-place.

dynamics::SVGD: The SVGD dynamics object specifying the kernel, step size (η), and batch size for the update.

NonparametricVI.KernelizedSteinDiscrepancy — Type

KernelizedSteinDiscrepancy <: Metric

A struct representing the Kernelized Stein Discrepancy (KSD) metric.

Fields

K::KernelFunctions.Kernel: The kernel function used in the KSD computation.

samplesize::Integer: the size of sample used to evaluate KSD

Description

This struct encapsulates the kernel function needed to compute the Kernelized Stein Discrepancy. The KSD is a measure of discrepancy between two probability distributions, and it relies on a kernel function to define the feature space in which the discrepancy is measured.

By holding the kernel function, this struct provides a convenient way to pass the necessary information for KSD computation to functions that track or evaluate the discrepancy between a particle distribution and a target distribution.

NonparametricVI.SVGD — Type

SVGD <: ParticleDynamics

Stein Variational Gradient Descent (SVGD) particle dynamics.

Fields

K::KernelFunctions.Kernel: The kernel function used to define the interaction between particles.

η: The step size or learning rate for updating particle positions.

batchsize: The number of particles to use in each update step (for mini-batching). If nothing, all particles are used.

Examples

using KernelFunctions

Define a squared exponential kernel

sqexp_kernel = SqExponentialKernel()

Create an SVGD dynamics object with a fixed step size and full batch

svgd_fullbatch = SVGD(K=sqexp_kernel, η=0.1, batchsize=nothing)

Create an SVGD dynamics object with a smaller step size and a batch size of 100

svgd_minibatch = SVGD(K=sqexp_kernel, η=0.05, batchsize=100)

NonparametricVI.SVGDInferenceReport — Type

SVGDInferenceReport <: InferenceReport

A mutable struct representing the report generated after running Stein Variational Gradient Descent (SVGD) inference.

Fields

metrics::Dict{String, Vector{Any}}: A dictionary storing various metrics tracked during the SVGD inference. Keys are metric names (strings), and values are vectors of metric values recorded at each iteration or relevant time step.

success::Bool: A boolean flag indicating whether the inference process completed successfully.

NonparametricVI.SVGDInferenceState — Type

SVGDInferenceState <: InferenceState

A mutable struct representing the internal state of the Stein Variational Gradient Descent (SVGD) inference algorithm.

Fields

ρ: A LogDensityProblem representing the target distribution's log-density function.

dynamics::SVGD: An SVGD object defining the dynamics used to update the particles.

model::Union{DynamicPPL.Model, Nothing}: The probabilistic model associated with the inference, or nothing if no model is provided.

Description

This struct encapsulates the state information required by the SVGD algorithm. It stores the target log-density, the SVGD dynamics, and optionally, a reference to the DynamicPPL model used. This state is passed between iterations of the SVGD algorithm to maintain and update the necessary information.